requirements.txt
face-recognition==1.3.0
multiprocess==0.70.16
pillow==11.0.0
requests==2.32.4
torch==2.7.1+cu128
numpy==2.1.2